requirements.txt
certifi==2018.8.24
chardet==3.0.4
elasticsearch==6.3.1
idna==2.7
influxdb==5.2.0
psutil==5.4.7
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
requests==2.19.1
six==1.11.0
urllib3==1.23
